requirements.txt
alembic==1.11.3
appdirs==1.4.4
appnope==0.1.3
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
behave==1.2.6
blinker==1.6.2
cattrs==23.1.2
certifi==2023.7.22
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
csvcubed==0.4.8
csvcubed-models==0.1.6
debugpy==1.6.7.post1
decorator==5.1.1
distlib==0.3.7
dnspython==2.4.2
email-validator==2.0.0.post2
executing==1.2.0
filelock==3.12.3
Flask==2.3.3
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
identify==2.5.13
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.3.1
jupyter_core==5.3.1
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
nodeenv==1.7.0
numpy==1.25.2
packaging==23.1
pandas==2.0.3
parse==1.19.1
parse-type==0.6.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
pre-commit==2.21.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.1
rdflib==7.0.0
requests==2.31.0
requests-cache==0.9.8
six==1.16.0
SQLAlchemy==2.0.20
stack-data==0.6.2
tornado==6.3.3
traitlets==5.9.0
treelib==1.7.0
typing_extensions==4.7.1
tzdata==2023.3
Unidecode==1.3.6
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.16
virtualenv==20.24.3
wcwidth==0.2.6
Werkzeug==2.3.7
WTForms==3.0.1
